+++ examples/acm-terraform-blog-part2/terraform/gke.tf
@@ -16,7 +16,7 @@
 
 module "enabled_google_apis" {
   source  = "terraform-google-modules/project-factory/google//modules/project_services"
-  version = "~> 16.0"
+  version = "~> 17.0"
 
   project_id                  = var.project
   disable_services_on_destroy = false
@@ -31,7 +31,7 @@ module "enabled_google_apis" {
 
 module "gke" {
   source             = "terraform-google-modules/kubernetes-engine/google"
-  version            = "~> 32.0"
+  version            = "~> 33.0"
   project_id         = module.enabled_google_apis.project_id
   name               = "sfl-acm-part2"
   region             = var.region