+++ modules/auth/versions.tf
@@ -21,11 +21,11 @@ terraform {
     google = {
       source = "hashicorp/google"
       # Avoid v4.49 and v4.50 for https://github.com/hashicorp/terraform-provider-google/issues/13507
-      version = ">= 4.47.0, != 4.49.0, != 4.50.0, < 6"
+      version = ">= 4.47.0, != 4.49.0, != 4.50.0, < 7"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-kubernetes-engine:auth/v33.0.0"
+    module_name = "blueprints/terraform/terraform-google-kubernetes-engine:auth/v33.1.0"
   }
 }